server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
gunicorn==21.2.0

# PDF 工具依赖
pypdf==4.0.0
pillow==10.2.0
pymupdf==1.23.8
